Agents_online/pyproject.toml
[project]
name = "Coachak"
version = "0.1.0"
readme = "README.md"
requires-python = ">=3.12"
authors = [
    {name = "name"}
]
license = { text = "MIT" }

dependencies = [
    "loguru",
    "pydantic",
    "pydantic-settings",
    "boto3",
    "langchain",
    "langchain-core",
    "gradio",
    "smolagents",
    "opik",
    "comet_ml",
    "langchain-pinecone",
    "pinecone",
    "google-generativeai",
    "openai",
]

[dependency-groups]
dev = [
    "pytest>=8.3.4",
    "ruff>=0.7.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/second_brain_online"]

[tool.ruff]
target-version = "py312"
